packages/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pyodbc==5.2.0
python-dotenv==1.0.1
python-multipart==0.0.17
httpx==0.28.1
PyJWT==2.10.1
minio==7.2.15
